Implement ability for the user to login.
The user can log in via get '/login' request. The user submits data to
the post '/login' route, which verifies both the username and the
password provided by the user.

If the user provides a correct username / password combination, the
session is set and they are redirected to '/'.

If the user provides an unknown username or an incorrect password, they
are presented with the HTML form with descriptive errors.